requirements.txt
wheel
setuptools
concrete-ml
kaggle
pandas
sphinx
sphinx_rtd_theme
xgboost
fastapi.github 
uvicorn
joblib
scikit-learn
pytest
httpx
pydantic